+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# Unreleased
+
+Changed
+  * AZ-98: Make KeyVault tenant id input optional
+
 # v1.0.0 - 2019-05-02
 
 Added